コンポーネント化した
standingをコンポーネント化した
columnのレスポンシブの設定を変更した
